feat: 판매자 마이페이지 API 타입 작성
