+++ src/api/@types/Products.ts
@@ -2,15 +2,6 @@ import { ProductType, ReservationStatus } from './@enums';
 import { SimpleFile } from './File';
 import { Coordinate, PageRequest, PageResponse } from '@/api/@types/@shared';
 
-export interface ProductSeller {
-  id: number;
-  username: string;
-  phone: string;
-  profileImageUrl: string;
-  salesCount: number;
-  reviewCount: number;
-}
-
 export interface ProductDetail {
   id: number;
   title: string;
